requirements.txt
# Core LangChain and LangGraph dependencies
langchain>=0.1.0
langchain-openai>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-google-genai>=2.0.0
langchain-anthropic>=0.1.0
langgraph>=0.2.0
wikipedia
# Search engine
tavily-python>=0.3.0

# LLM providers
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.4.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Progress tracking and utilities
tqdm>=4.65.0

# Environment and configuration
python-dotenv>=1.0.0

# Type hints and annotations
typing-extensions>=4.8.0

# Vector database and embeddings
chromadb>=0.4.24
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# Text processing
scikit-learn>=1.3.0

# JSON handling (built-in, but listed for clarity)
# json - built-in module

# Regular expressions (built-in, but listed for clarity)
# re - built-in module

# File system operations (built-in, but listed for clarity)
# os - built-in module
# time - built-in module
# random - built-in module

fastapi
uvicorn
uqlm